requirements.txt
pytesseract==0.3.10
google-cloud-vision==3.4.4
pydantic==2.5.0
spacy==3.7.2
streamlit==1.28.1
Pillow==10.1.0
pdf2image==1.17.0
opencv-python==4.8.1.78
numpy==1.24.3
pandas==2.1.3
pdfplumber==0.10.3
scikit-image==0.21.0
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
